requirements.txt
fastapi
uvicorn
pillow
pillow-heif
python-multipart
torch
torchvision
opencv-python
timm